+++ swiyu-trust-management-ui/src/app/pages/trust-onboarding-task-list/trust-onboarding-task-list.component.ts
@@ -1,5 +1,5 @@
 import {DatePipe} from '@angular/common';
-import {AfterViewInit, Component, ViewChild, inject} from '@angular/core';
+import {AfterViewInit, Component, inject, ViewChild} from '@angular/core';
 import {takeUntilDestroyed} from '@angular/core/rxjs-interop';
 import {FormBuilder, FormControl, FormGroup, ReactiveFormsModule} from '@angular/forms';
 import {MatButtonModule} from '@angular/material/button';