+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 3.31.6
+
+### Changed
+
+- fix user role handling for int (preview) stages
+- 
+## 3.31.5
+
+### Changed
+
+- enabled default oblique spinner and error interceptor
+
 ## 3.31.4
 
 ### Changed